refactor: modularize data enrichment
